backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ML/Embeddings
torch==2.3.1+cpu
sentence-transformers==3.0.1
faiss-cpu==1.9.0

# Data processing
pandas==2.2.2
pyarrow==16.1.0

# AWS/Storage
boto3==1.34.131

# Utils
python-dotenv==1.0.1
pydantic>=2.9.0,<3.0.0

# LLM
google-genai==1.52.0